Move componentes para pasta Contratos
