+++ src/components/Contratos/TableContrato/style.scss
