+++ src/pages/ContratosContinuos/index.jsx
@@ -3,10 +3,10 @@ import Page from "../../components/Page";
 import Container from "../../components/Container";
 import { TabView, TabPanel } from "primereact/tabview";
 import { Accordion, AccordionTab } from "primereact/accordion";
-import { TableContrato } from "../../components/TableContrato";
+import { TableContrato } from "../../components/Contratos/TableContrato";
 import { getContratos } from "../../service/Contratos.service";
 import "./style.scss";
-import { BuscaContratosForm } from "../../components/Coad/BuscaContratosForm";
+import { BuscaContratosForm } from "../../components/Contratos/BuscaContratosForm";
 import { getUsuario } from "../../service/auth.service";
 import { getUrlParams } from "../../utils/params";
 import { Button, ButtonGroup } from "reactstrap";
