+++ src/pages/PainelSelecao/index.jsx
@@ -1,11 +1,11 @@
 import React, { Component } from "react";
 import Page from "../../components/Page";
-import CardSection from "../../components/CardSection";
+import CardSection from "../../components/Contratos/CardSection";
 import IconeEquipamento from "../../assets/images/icon-equipamento.svg";
 import IconeServico from "../../assets/images/icon-servicos.svg";
 import IconeMapa from "../../assets/images/icon-mapa.svg";
-import CardEquipamento from "../../components/CardEquipamento";
-import CardServico from "../../components/CardServico";
+import CardEquipamento from "../../components/Contratos/CardEquipamento";
+import CardServico from "../../components/Contratos/CardServico";
 import { Button, ButtonGroup } from "reactstrap";
 import { redirect } from "../../utils/redirect";
